feat(Home): update service items with detailed descriptions and improve layout feat(ChangeLanguage): add language selection screen with options for English and Arabic refactor(MyAccount): convert to StatefulWidget, enhance logout confirmation dialog, and improve loading state handling chore(Podfile.lock): add missing pod dependencies for macOS chore(project.pbxproj): update project settings to include new pod frameworks chore(contents.xcworkspacedata): include Pods project in workspace references

+++ lib/presentation/screens/user/my_account.dart
@@ -5,8 +5,16 @@ import 'package:sanam_laundry/data/index.dart';
 import 'package:sanam_laundry/presentation/index.dart';
 import 'package:sanam_laundry/providers/auth.dart';
 
-class MyAccount extends StatelessWidget {
+class MyAccount extends StatefulWidget {
   const MyAccount({super.key});
+
+  @override
+  State<MyAccount> createState() => _MyAccountState();
+}
+
+class _MyAccountState extends State<MyAccount> {
+  bool isLoading = false;
+  final AuthRepository _authRepository = AuthRepository();
 
   @override
   Widget build(BuildContext context) {
@@ -43,7 +51,7 @@ class MyAccount extends StatelessWidget {
         icon: Icons.translate,
         title: Common.language,
         onTap: () {
-          // context.navigate(AppRoutes.language);
+          context.navigate(AppRoutes.language);
         },
       ),
       _AccountOption(
@@ -81,22 +89,60 @@ class MyAccount extends StatelessWidget {
             context,
             title: Common.deleteAccount,
             borderColor: AppColors.primary,
-            borderWidth: 4,
+            borderWidth: 5,
+            dismissible: false,
             borderRadius: Dimens.radiusL,
-            imageSize: 150,
             content: AppText(
               Common.doYouWantToDeleteAccount,
               maxLines: 3,
               textAlign: TextAlign.center,
+              style: context.textTheme.bodyLarge?.copyWith(
+                fontWeight: FontWeight.bold,
+              ),
             ),
-            primaryButtonText: Common.deleteAccount,
-            onPrimaryPressed: () => {
-              AuthService.removeToken(),
-              context.replacePage(AppRoutes.getStarted),
+            showTwoPrimaryButtons: true,
+            primaryButtonText: Common.yes,
+            secondaryButtonText: Common.no,
+            onPrimaryPressed: () async {
+              setState(() {
+                isLoading = true;
+              });
+              context.back();
+              final currentContext = context;
+              final isDeleted = await _authRepository.deleteAccount();
+              if (isDeleted != null) {
+                if (!currentContext.mounted) return;
+                setState(() {
+                  isLoading = false;
+                });
+                AuthService.removeToken();
+                AppDialog.show(
+                  context,
+                  borderColor: AppColors.primary,
+                  borderWidth: 5,
+                  dismissible: false,
+                  borderRadius: Dimens.radiusL,
+                  content: AppText(
+                    Common.yourAccountHasBeenDeleted,
+                    maxLines: 3,
+                    textAlign: TextAlign.center,
+                    style: context.textTheme.bodyLarge?.copyWith(
+                      fontWeight: FontWeight.bold,
+                    ),
+                  ),
+                  primaryButtonText: Common.signUpAgain,
+                  onPrimaryPressed: () async {
+                    context.replacePage(AppRoutes.getStarted);
+                  },
+                  crossAxisAlignment: CrossAxisAlignment.center,
+                  insetPadding: EdgeInsets.all(Dimens.spacingMLarge),
+                );
+              }
             },
+            onSecondaryPressed: () => {context.back()},
             backgroundColor: AppColors.lightWhite,
             crossAxisAlignment: CrossAxisAlignment.center,
-            insetPadding: EdgeInsets.all(Dimens.spacingXXL),
+            insetPadding: EdgeInsets.all(Dimens.spacingMLarge),
           );
         },
       ),
@@ -106,39 +152,97 @@ class MyAccount extends StatelessWidget {
       scrollable: true,
       safeArea: false,
       padding: EdgeInsets.zero,
-      child: Column(
+      child: Stack(
         children: [
-          _buildHeader(context),
-          Padding(
-            padding: const EdgeInsets.symmetric(
-              horizontal: Dimens.screenMarginVertical,
-            ),
-            child: Column(
-              children: options
-                  .map(
-                    (e) => ListTile(
-                      leading: AppIcon(icon: e.icon, color: AppColors.primary),
-                      title: AppText(e.title),
-                      trailing: const AppIcon(
-                        icon: Icons.arrow_forward_ios,
-                        size: 14,
-                        color: Colors.grey,
-                      ),
-                      onTap: e.onTap,
+          Positioned(
+            bottom: context.w(0.8),
+            left: 0,
+            right: 0,
+            child: IgnorePointer(
+              child: Opacity(
+                opacity: 0.07,
+                child: Transform.rotate(
+                  angle:
+                      -1.5, // in radians → about -30° rotation (adjust as needed)
+                  alignment: Alignment.centerRight, // rotate around right side
+                  child: Align(
+                    alignment:
+                        Alignment.centerRight, // ensure it starts from right
+                    child: AppImage(
+                      path: AppAssets.watermark,
+                      width: context.screenWidth,
+                      height: context.h(0.60),
+                      fit: BoxFit.cover,
                     ),
-                  )
-                  .toList(),
+                  ),
+                ),
+              ),
             ),
           ),
-          Padding(
-            padding: const EdgeInsets.symmetric(vertical: Dimens.spacingM),
-            child: AppButton(
-              title: Common.logout,
-              onPressed: () {
-                AuthService.removeToken();
-                context.replacePage(AppRoutes.getStarted);
-              },
-            ),
+
+          Column(
+            children: [
+              _buildHeader(context),
+              Padding(
+                padding: const EdgeInsets.symmetric(
+                  horizontal: Dimens.screenMarginVertical,
+                ),
+                child: Column(
+                  children: options
+                      .map(
+                        (e) => ListTile(
+                          leading: AppIcon(
+                            icon: e.icon,
+                            color: AppColors.primary,
+                          ),
+                          title: AppText(e.title),
+                          trailing: const AppIcon(
+                            icon: Icons.arrow_forward_ios,
+                            size: 14,
+                            color: Colors.grey,
+                          ),
+                          onTap: e.onTap,
+                        ),
+                      )
+                      .toList(),
+                ),
+              ),
+              Padding(
+                padding: const EdgeInsets.symmetric(vertical: Dimens.spacingM),
+                child: AppButton(
+                  isLoading: isLoading,
+                  title: Common.logout,
+                  onPressed: () {
+                    AppDialog.show(
+                      context,
+                      title: Common.logout,
+                      borderColor: AppColors.primary,
+                      borderWidth: 5,
+                      dismissible: false,
+                      borderRadius: Dimens.radiusL,
+                      content: AppText(
+                        Common.areYouSureYouWantToLogout,
+                        maxLines: 3,
+                        textAlign: TextAlign.center,
+                        style: context.textTheme.bodyLarge?.copyWith(
+                          fontWeight: FontWeight.bold,
+                        ),
+                      ),
+                      primaryButtonText: Common.yes,
+                      secondaryButtonText: Common.no,
+                      showTwoPrimaryButtons: true,
+                      onSecondaryPressed: () => {context.back()},
+                      onPrimaryPressed: () async {
+                        context.replacePage(AppRoutes.getStarted);
+                        _authRepository.logout();
+                      },
+                      crossAxisAlignment: CrossAxisAlignment.center,
+                      insetPadding: EdgeInsets.all(Dimens.spacingMLarge),
+                    );
+                  },
+                ),
+              ),
+            ],
           ),
         ],
       ),
@@ -170,14 +274,11 @@ class MyAccount extends StatelessWidget {
             child: Consumer<AuthProvider>(
               builder: (context, auth, child) {
                 final profileImage = auth.user?.profileImage ?? AppAssets.user;
-                return CircleAvatar(
-                  radius: 42,
-                  child: AppImage(
-                    path: profileImage,
-                    height: context.h(0.2),
-                    width: context.screenWidth,
-                    fit: BoxFit.contain,
-                  ),
+                return AppImage(
+                  path: profileImage,
+                  height: 90,
+                  width: 90,
+                  isCircular: true,
                 );
               },
             ),
